+++ packages/ragbits-chat/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- remove HeroUI Pro components (#557)
+
 ## 0.18.0 (2025-05-22)
 
 ### Changed